src/python/czo_yaml/requirements.txt
Cython==0.19.1
Django==1.6.2
Flask==0.10.1
Jinja2==2.6
MDP==3.3
Mezzanine==3.0.9
PIL==1.1.7
PySAL==1.5.0
PySide==1.1.2
PyYAML==3.10
Pygments==1.6
SQLAlchemy==0.8.1
Sphinx==1.1.3
Werkzeug==0.9.1
astropy==0.2.3
atom==0.2.3
binstar-client==0.1.0
biopython==1.61
bitarray==0.8.1
bleach==1.4
boto==2.9.6
casuarius==1.1
chaco==4.2.1
conda==1.7.2
cubes==0.10.2
distribute==0.6.45
docutils==0.10
enable==4.2.1
enaml==0.7.6
filebrowser-safe==0.3.2
future==0.9.0
gevent==0.13.8
gevent-websocket==0.3.6
gevent-zeromq==0.2.2
grappelli-safe==0.3.7
greenlet==0.4.1
grin==1.2.1
h5py==2.1.1
html5lib==1.0b3
ipython==0.13.2
itsdangerous==0.21
keyring==1.4
llvmmath==0.1
llvmpy==0.11.3
lxml==3.2.1
matplotlib==1.2.1
menuinst==1.0.1
meta==development
networkx==1.7
nltk==2.0.4
nose==1.3.0
numba==0.9.0
numexpr==2.0.1
numpy==1.7.1
oauthlib==0.6.1
pandas==0.11.0
pep8==1.4.5
ply==3.4
psutil==0.7.1
py==1.4.14
pycosat==0.6.0
pycparser==2.09.1
pycrypto==2.6
pyface==4.2.1
pyflakes==0.7.2
pyparsing==1.5.6
pyreadline==2.0-dev1
pytest==2.3.5
python-dateutil==2.1
pytz==2014.2
pywin32==218.4
pyzmq==2.2.0.1
requests==2.2.1
requests-oauthlib==0.4.0
rope==0.9.4
scikit-image==0.8.2
scikit-learn==0.13.1
scipy==0.12.0
six==1.6.1
spyder==2.2.0
statsmodels==0.4.3
sympy==0.7.2
tables==2.4.0
tornado==3.1
traits==4.2.1
traitsui==4.2.1
tzlocal==1.0
wsgiref==0.1.2
xlrd==0.9.2
xlwt==0.7.5
